Factor out GenerateAutoloadReferenceFile.
