+++ src/ComposerScaffoldCommand.php
@@ -29,8 +29,6 @@ class ComposerScaffoldCommand extends BaseCommand {
   protected function execute(InputInterface $input, OutputInterface $output) {
     $handler = new Handler($this->getComposer(), $this->getIO());
     $handler->scaffold();
-    // Generate the autoload.php file after generating the scaffold files.
-    $handler->generateAutoload();
   }
 
 }
