+++ src/Handler.php
@@ -10,7 +10,6 @@ use Composer\Composer;
 use Composer\EventDispatcher\EventDispatcher;
 use Composer\IO\IOInterface;
 use Composer\Util\Filesystem;
-use Symfony\Component\Filesystem\Filesystem as SymfonyFilesystem;
 use Grasmash\ComposerScaffold\Operations\ScaffoldCopyOp;
 use Grasmash\ComposerScaffold\Operations\ScaffoldSymlinkOp;
 use Grasmash\ComposerScaffold\Operations\ScaffoldReplaceOp;
@@ -136,70 +135,11 @@ class Handler {
     // Generate an autoload file in the document root that includes
     // the autoload.php file in the vendor directory, wherever that is.
     // Drupal requires this in order to easily locate relocated vendor dirs.
-    $this->generateAutoload();
+    $generator = new GenerateAutoloadReferenceFile($this->composer);
+    $generator->generateAutoload($this->getWebRoot());
 
     // Call post-scaffold scripts.
     $dispatcher->dispatch(self::POST_COMPOSER_SCAFFOLD_CMD);
-  }
-
-  /**
-   * Generate the autoload file at the project root.
-   *
-   * Include the autoload file that Composer generated.
-   */
-  public function generateAutoload() {
-    $vendorPath = $this->getVendorPath();
-    $webroot = $this->getWebRoot();
-
-    // Calculate the relative path from the webroot (location of the project
-    // autoload.php) to the vendor directory.
-    $fs = new SymfonyFilesystem();
-    $relativeVendorPath = $fs->makePathRelative($vendorPath, realpath($webroot));
-
-    $fs->dumpFile($webroot . "/autoload.php", $this->autoLoadContents($relativeVendorPath));
-  }
-
-  /**
-   * Build the contents of the autoload file.
-   *
-   * @return string
-   *   Return the contents for the autoload.php.
-   */
-  protected function autoLoadContents(string $relativeVendorPath) : string {
-    $relativeVendorPath = rtrim($relativeVendorPath, '/');
-
-    return <<<EOF
-<?php
-
-/**
- * @file
- * Includes the autoloader created by Composer.
- *
- * This file was generated by composer-scaffold.
- *.
- * @see composer.json
- * @see index.php
- * @see core/install.php
- * @see core/rebuild.php
- * @see core/modules/statistics/statistics.php
- */
-
-return require __DIR__ . '/$relativeVendorPath/autoload.php';
-
-EOF;
-  }
-
-  /**
-   * Get the path to the 'vendor' directory.
-   *
-   * @return string
-   *   The file path of the vendor directory.
-   */
-  public function getVendorPath() {
-    $vendorDir = $this->composer->getConfig()->get('vendor-dir');
-    $filesystem = new Filesystem();
-    $filesystem->ensureDirectoryExists($vendorDir);
-    return $filesystem->normalizePath(realpath($vendorDir));
   }
 
   /**
@@ -330,8 +270,7 @@ EOF;
     ];
     $allowed_packages = $this->recursiveGetAllowedPackages($options['allowed-packages']);
 
-    // Add root package at the end so that it overrides all the preceding
-    // package.
+    // Add root package at the end so that it overrides all the preceding package.
     $root_package = $this->composer->getPackage();
     $allowed_packages[$root_package->getName()] = $root_package;
